Add deploy feature to build.sh

+++ build.sh
@@ -7,16 +7,45 @@ if [[ PYVER -ne 3 ]];then
   exit 1
 fi
 
-pip install -r requirements.txt
-WGET_VER=$(curl http://tbears.icon.foundation.s3-website.ap-northeast-2.amazonaws.com/earlgrey/VERSION)
-pip install --force-reinstall "http://tbears.icon.foundation.s3-website.ap-northeast-2.amazonaws.com/earlgrey/earlgrey-${WGET_VER}-py3-none-any.whl"
-rm -rf earlgrey*
+if [[ ("$1" = "test" && "$2" != "--ignore-test") || ("$1" = "build") || ("$1" = "deploy") ]]; then
+  pip install -r requirements.txt
 
-WGET_VER=$(curl http://tbears.icon.foundation.s3-website.ap-northeast-2.amazonaws.com/iconcommons/VERSION)
-pip install --force-reinstall "http://tbears.icon.foundation.s3-website.ap-northeast-2.amazonaws.com/iconcommons/iconcommons-${WGET_VER}-py3-none-any.whl"
-rm -rf iconcommons*
+  WGET_VER=$(curl http://tbears.icon.foundation.s3-website.ap-northeast-2.amazonaws.com/earlgrey/VERSION)
+  pip install --force-reinstall "http://tbears.icon.foundation.s3-website.ap-northeast-2.amazonaws.com/earlgrey/earlgrey-${WGET_VER}-py3-none-any.whl"
 
+  WGET_VER=$(curl http://tbears.icon.foundation.s3-website.ap-northeast-2.amazonaws.com/iconcommons/VERSION)
+  pip install --force-reinstall "http://tbears.icon.foundation.s3-website.ap-northeast-2.amazonaws.com/iconcommons/iconcommons-${WGET_VER}-py3-none-any.whl"
 
-pip install wheel
-rm -rf build dist *.egg-info
-python setup.py bdist_wheel
+  if [[ "$2" != "--ignore-test" ]]; then
+    python -m unittest
+  fi
+
+  if [ "$1" = "build" ] || [ "$1" = "deploy" ]; then
+    pip install wheel
+    rm -rf build dist *.egg-info
+    python setup.py bdist_wheel
+
+    if [ "$1" = "deploy" ]; then
+      VER=$(cat VERSION)
+
+      if [[ -z "${AWS_ACCESS_KEY_ID}" || -z "${AWS_SECRET_ACCESS_KEY}" ]]; then
+        echo "Error: AWS keys should be in your environment"
+        exit 1
+      fi
+
+      pip install awscli
+      aws s3 cp VERSION s3://tbears.icon.foundation/iconrpcserver/ --acl public-read
+      aws s3 cp dist/*$VER*.whl s3://tbears.icon.foundation/iconrpcserver/ --acl public-read
+
+    fi
+  fi
+
+else
+  echo "Usage: build.sh [test|build|deploy]"
+  echo "  test: run test"
+  echo "  build: run test and build"
+  echo "  build --ignore-test: run build"
+  echo "  deploy: run test, build and deploy to s3"
+  echo "  deploy --ignore-test: run build and deploy to s3"
+  exit 1
+fi
